imageAnalyzer/requirements.txt
# requirements.txt

requests
flask
flask_restful
google.cloud.vision